Use common util func to create sounds, users, packs when testing
