+++ sounds/tests.py
@@ -47,6 +47,7 @@ from sounds.views import get_sound_of_the_day_id
 from accounts.models import EmailPreferenceType
 from utils.encryption import encrypt
 from utils.tags import clean_and_split_tags
+from utils.test import create_user_and_sounds
 from utils.cache import get_template_cache_key
 
 from bs4 import BeautifulSoup
@@ -244,38 +245,6 @@ class CommentSoundsTestCase(TestCase):
         sound = Sound.objects.get(id=19)
         self.assertEqual(current_num_comments, sound.num_comments)
         self.assertEqual(sound.is_index_dirty, True)
-
-
-sound_counter = count()
-
-
-def create_user_and_sounds(num_sounds=1, num_packs=0, user=None, count_offset=0, tags=None,
-                           processing_state='PE', moderation_state='PE'):
-    count_offset = count_offset + next(sound_counter)
-    if user is None:
-        user = User.objects.create_user("testuser", password="testpass", email='email@freesound.org')
-    packs = list()
-    for i in range(0, num_packs):
-        pack = Pack.objects.create(user=user, name="Test pack %i" % (i + count_offset))
-        packs.append(pack)
-    sounds = list()
-    for i in range(0, num_sounds):
-        pack = None
-        if packs:
-            pack = packs[i % len(packs)]
-        sound = Sound.objects.create(user=user,
-                                     original_filename="Test sound %i" % (i + count_offset),
-                                     base_filename_slug="test_sound_%i" % (i + count_offset),
-                                     license=License.objects.all()[0],
-                                     pack=pack,
-                                     md5="fakemd5_%i" % (i + count_offset),
-                                     processing_state=processing_state,
-                                     moderation_state=moderation_state)
-
-        if tags is not None:
-            sound.set_tags(clean_and_split_tags(tags))
-        sounds.append(sound)
-    return user, packs, sounds
 
 
 class ChangeSoundOwnerTestCase(TestCase):
