+++ utils/test.py
@@ -19,6 +19,12 @@
 #
 
 import os
+from itertools import count
+
+from django.contrib.auth.models import User
+
+from sounds.models import Sound, Pack, License
+from utils.tags import clean_and_split_tags
 
 
 def create_test_files(filenames, directory, n_bytes=1024):
@@ -32,3 +38,50 @@ def create_test_files(filenames, directory, n_bytes=1024):
         f = open(os.path.join(directory, filename), 'a')
         f.write(os.urandom(n_bytes))
         f.close()
+
+
+sound_counter = count()  # Used in create_user_and_sounds to avoid repeating sound names
+
+
+def create_user_and_sounds(num_sounds=1, num_packs=0, user=None, count_offset=0, tags=None,
+                           processing_state='PE', moderation_state='PE'):
+    """
+    Creates User, Sound and Pack objects useful for testing. A 'sound_counter' is used to make sound names unique as
+    well as other fields like md5.
+    NOTE: creating sounds requires License objects to exist in DB. Do that by making sure your test case loads
+    'initial_data' fixture, i.e. "fixtures = ['initial_data']".
+    :param num_sounds: N sounds to generate
+    :param num_packs: N packs in which the sounds above will be grouped
+    :param user: user owner of the created sounds (if not provided, a new user will be created)
+    :param count_offset: start counting sounds at X
+    :param tags: tags added to the sounds (all sounds will have the same tags)
+    :param processing_state: processing state of the created sounds
+    :param moderation_state: moderation state of the created sounds
+    :return: tuple with (User object, list of Pack objcts, list of Sound objects)
+    """
+    count_offset = count_offset + next(sound_counter)
+    if user is None:
+        user = User.objects.create_user("testuser", password="testpass", email='email@freesound.org')
+    packs = list()
+    for i in range(0, num_packs):
+        pack = Pack.objects.create(user=user, name="Test pack %i" % (i + count_offset))
+        packs.append(pack)
+    sounds = list()
+    for i in range(0, num_sounds):
+        pack = None
+        if packs:
+            pack = packs[i % len(packs)]
+        sound = Sound.objects.create(user=user,
+                                     original_filename="Test sound %i" % (i + count_offset),
+                                     base_filename_slug="test_sound_%i" % (i + count_offset),
+                                     license=License.objects.all()[0],
+                                     pack=pack,
+                                     md5="fakemd5_%i" % (i + count_offset),
+                                     processing_state=processing_state,
+                                     moderation_state=moderation_state)
+
+        if tags is not None:
+            sound.set_tags(clean_and_split_tags(tags))
+        sounds.append(sound)
+    return user, packs, sounds
+
